+++ bika/lims/browser/analysisrequest/analysisrequests.py
@@ -730,11 +730,8 @@ class AnalysisRequestsView(BikaListingView):
                        src='%s/++resource++bika.lims.images/unassigned.png'/>"\
                        % (t(_("Unassigned")), self.portal_url),
              'contentFilter': {'worksheetanalysis_review_state': 'unassigned',
-                               'review_state': ('sample_registered',
-                                                'sample_prep',
-                                                'to_be_sampled',
-                                                'sample_due',
-                                                'sample_received',
+                               'cancellation_state': 'active',
+                               'review_state': ('sample_received',
                                                 'attachment_due',),
                                'sort_on': 'Created',
                                'sort_order': 'reverse'},