+++ CHANGES.rst
@@ -35,7 +35,7 @@ Changelog
 
 **Changed**
 
-- Changed Unassigned filter
+- #627 Unassigned filter on Analysis Requests view
 - #621 Change Errors to Warnings when importing instrument results
 
 **Fixed**